Decouple user input in remaining game states
