+++ include/minefield/select_mines.h
@@ -5,7 +5,7 @@
 
 NextState selectMines(GameContext &context);
 NextState selectGuesses(GameContext &context);
-void manualMineSelection(int selectionSize, std::vector<unsigned int> &minesVector, const std::vector<unsigned int> &availableCells);
+void manualMineSelection(int selectionSize, std::vector<unsigned int> &minesVector, const std::vector<unsigned int> &availableCells, GetInputFn<unsigned int> getInput);
 void randomMineSelection(int selectionSize, std::vector<unsigned int> &randomSelection, const std::vector<unsigned int> &availableCells, const RandomGenerationFn &randomGenerator);
 void printMineSelection(const std::vector<int> &mines, const std::string &message = "");
 unsigned int getRivalsMaxMineCount(const Player &player, const std::vector<Player> &players);
